Не спрашивайте

+++ 02.Memory_Overcommit/WSL_codes/memory_overcommit_wsl.c
@@ -2,7 +2,7 @@
 #include <stdlib.h>
 #include <unistd.h>
 
-typedef unsigned int type_size;
+typedef unsigned long long type_size;
 typedef long long type_array;
 #define INTERVAL 0.8
 #define LIMIT 2560
@@ -32,7 +32,7 @@ int main(int argc, char* argv[]) {
     int flag;
     type_size size = 1;
     size, flag = get_data(argv, &size, &flag);
-    printf("size = %u, flag = %i\n", size, flag);
+    printf("size = %llu, flag = %i\n", size, flag);
     type_array *array = malloc(sizeof(type_array)*size);
     overcommit(array, size, flag);
     printf("STOP!!!\n");
